move vertex types into own files
