+++ src/providers/google-vertex-ai/chatComplete.ts
@@ -1,3 +1,6 @@
+// Docs for REST API
+// https://cloud.google.com/vertex-ai/generative-ai/docs/multimodal/send-multimodal-prompts#gemini-send-multimodal-samples-drest
+
 import { GOOGLE_VERTEX_AI } from '../../globals';
 import { ContentType, Message, Params } from '../../types/requestBody';
 import {
@@ -9,31 +12,11 @@ import {
   generateErrorResponse,
   generateInvalidProviderResponseError,
 } from '../utils';
-
-const transformGenerationConfig = (params: Params) => {
-  const generationConfig: Record<string, any> = {};
-  if (params['temperature']) {
-    generationConfig['temperature'] = params['temperature'];
-  }
-  if (params['top_p']) {
-    generationConfig['topP'] = params['top_p'];
-  }
-  if (params['top_k']) {
-    generationConfig['topK'] = params['top_k'];
-  }
-  if (params['max_tokens']) {
-    generationConfig['maxOutputTokens'] = params['max_tokens'];
-  }
-  if (params['stop']) {
-    generationConfig['stopSequences'] = params['stop'];
-  }
-  return generationConfig;
-};
-
-// TODOS: this configuration does not enforce the maximum token limit for the input parameter. If you want to enforce this, you might need to add a custom validation function or a max property to the ParameterConfig interface, and then use it in the input configuration. However, this might be complex because the token count is not a simple length check, but depends on the specific tokenization method used by the model.
-
-// Docs for REST API
-// https://cloud.google.com/vertex-ai/generative-ai/docs/multimodal/send-multimodal-prompts#gemini-send-multimodal-samples-drest
+import { transformGenerationConfig } from './transformGenerationConfig';
+import type {
+  GoogleErrorResponse,
+  GoogleGenerateContentResponse,
+} from './types';
 
 export const GoogleChatCompleteConfig: ProviderConfig = {
   // https://cloud.google.com/vertex-ai/generative-ai/docs/learn/model-versioning#gemini-model-versions
@@ -170,51 +153,6 @@ export const GoogleChatCompleteConfig: ProviderConfig = {
     },
   },
 };
-
-export interface GoogleErrorResponse {
-  error: {
-    code: number;
-    message: string;
-    status: string;
-    details: Array<Record<string, any>>;
-  };
-}
-
-interface GoogleGenerateFunctionCall {
-  name: string;
-  args: Record<string, any>;
-}
-
-interface GoogleGenerateContentResponse {
-  candidates: {
-    content: {
-      parts: {
-        text?: string;
-        functionCall?: GoogleGenerateFunctionCall;
-      }[];
-    };
-    finishReason: string;
-    index: 0;
-    safetyRatings: {
-      category: string;
-      probability: string;
-    }[];
-  }[];
-  promptFeedback: {
-    safetyRatings: {
-      category: string;
-      probability: string;
-      probabilityScore: number;
-      severity: string;
-      severityScore: number;
-    }[];
-  };
-  usageMetadata: {
-    promptTokenCount: number;
-    candidatesTokenCount: number;
-    totalTokenCount: number;
-  };
-}
 
 export const GoogleChatCompleteResponseTransform: (
   response:
