+++ src/providers/google-vertex-ai/transformGenerationConfig.ts
@@ -1,0 +1,25 @@
+import { Params } from '../../types/requestBody';
+
+/**
+ * @see https://cloud.google.com/vertex-ai/generative-ai/docs/model-reference/gemini#request_body
+ */
+export function transformGenerationConfig(params: Params) {
+  const generationConfig: Record<string, any> = {};
+  if (params['temperature']) {
+    generationConfig['temperature'] = params['temperature'];
+  }
+  if (params['top_p']) {
+    generationConfig['topP'] = params['top_p'];
+  }
+  if (params['top_k']) {
+    generationConfig['topK'] = params['top_k'];
+  }
+  if (params['max_tokens']) {
+    generationConfig['maxOutputTokens'] = params['max_tokens'];
+  }
+  if (params['stop']) {
+    generationConfig['stopSequences'] = params['stop'];
+  }
+
+  return generationConfig;
+}
