+++ src/providers/google-vertex-ai/types.ts
@@ -1,0 +1,44 @@
+export interface GoogleErrorResponse {
+  error: {
+    code: number;
+    message: string;
+    status: string;
+    details: Array<Record<string, any>>;
+  };
+}
+
+export interface GoogleGenerateFunctionCall {
+  name: string;
+  args: Record<string, any>;
+}
+
+export interface GoogleGenerateContentResponse {
+  candidates: {
+    content: {
+      parts: {
+        text?: string;
+        functionCall?: GoogleGenerateFunctionCall;
+      }[];
+    };
+    finishReason: string;
+    index: 0;
+    safetyRatings: {
+      category: string;
+      probability: string;
+    }[];
+  }[];
+  promptFeedback: {
+    safetyRatings: {
+      category: string;
+      probability: string;
+      probabilityScore: number;
+      severity: string;
+      severityScore: number;
+    }[];
+  };
+  usageMetadata: {
+    promptTokenCount: number;
+    candidatesTokenCount: number;
+    totalTokenCount: number;
+  };
+}
